Search components separated, Header improved
